+++ crates/mcp/src/mcp.rs
@@ -1,4 +1,4 @@
-use crate::tools::action_tools::SyncWorkspaceTool;
+use crate::tools::action_tools::{SyncProjectsTool, SyncWorkspaceTool};
 use crate::tools::project_tools::{GetProjectTool, GetProjectsTool};
 use crate::tools::task_tools::{GetTaskTool, GetTasksTool};
 use crate::tools::vcs_tools::GetTouchedFiles;
@@ -49,6 +49,11 @@ impl ServerHandler for MoonMcpHandler {
             MoonTools::GetTaskTool(inner) => inner.call_tool(&self.workspace_graph),
             MoonTools::GetTasksTool(inner) => inner.call_tool(&self.workspace_graph),
             MoonTools::GetTouchedFiles(inner) => inner.call_tool(&self.app_context).await,
+            MoonTools::SyncProjectsTool(inner) => {
+                inner
+                    .call_tool(&self.app_context, &self.workspace_graph)
+                    .await
+            }
             MoonTools::SyncWorkspaceTool(inner) => {
                 inner
                     .call_tool(&self.app_context, &self.workspace_graph)
@@ -102,6 +107,7 @@ tool_box!(
         GetTaskTool,
         GetTasksTool,
         GetTouchedFiles,
+        SyncProjectsTool,
         SyncWorkspaceTool
     ]
 );